refactor: move 'store' directory into 'app' directory
